+++ CHANGELOG.md
@@ -7,11 +7,20 @@ This document outlines major changes between releases.
 ### Added
 
 ### Changed
-- Disable URI path escaping for presigned URLs (#1046)
 
 ### Fixed
 
 ### Updated
+
+### Removed
+
+## [0.34.0] - 2024-12-30
+
+### Changed
+- Disable URI path escaping for presigned URLs (#1046)
+
+### Updated
+- golang.org/x/crypto dependency from 0.26.0 to 0.31.0 (#1044)
 
 ### Removed
 - Command generate-presigned-url from authmate (#1048)
@@ -677,4 +686,5 @@ releases.
 [0.31.1]: https://github.com/nspcc-dev/neofs-s3-gw/compare/v0.31.0...v0.31.1
 [0.32.0]: https://github.com/nspcc-dev/neofs-s3-gw/compare/v0.31.1...v0.32.0
 [0.33.0]: https://github.com/nspcc-dev/neofs-s3-gw/compare/v0.32.0...v0.33.0
-[Unreleased]: https://github.com/nspcc-dev/neofs-s3-gw/compare/v0.33.0...master
+[0.34.0]: https://github.com/nspcc-dev/neofs-s3-gw/compare/v0.33.0...v0.34.0
+[Unreleased]: https://github.com/nspcc-dev/neofs-s3-gw/compare/v0.34.0...master